refactor: Message 분리
